+++ platform/flowglad-next/src/app/api/hosted-billing/request-magic-link/route.ts
@@ -6,14 +6,17 @@ import {
 } from '@/utils/hostedBillingApiHelpers'
 import { adminTransaction } from '@/db/adminTransaction'
 import {
+  mapCustomerEmailToStackAuthHostedBillingUserId,
   selectCustomers,
-  updateCustomer,
 } from '@/db/tableMethods/customerMethods'
 import { hostedBillingStackServerApp } from '@/stack'
 import { logger } from '@/utils/logger'
-import { trace, SpanStatusCode, context } from '@opentelemetry/api'
-import { ServerUser } from '@stackframe/stack'
-import core from '@/utils/core'
+import { trace, SpanStatusCode } from '@opentelemetry/api'
+import {
+  createUserAndSendMagicLink,
+  sendEmailToExistingUser,
+} from './sendMagicLinkHandlers'
+import { Customer } from '@/db/schema/customers'
 
 const requestSchema = z.object({
   organizationId: z.string(),
@@ -21,187 +24,75 @@ const requestSchema = z.object({
   customerExternalId: z.string(),
 })
 
-const sendVerificationEmailToUser = async (
-  user: ServerUser,
-  params: {
-    organizationId: string
-    customerExternalId: string
-  }
-) => {
-  const contactChannels = await user.listContactChannels()
-  const primaryContactChannel = contactChannels.find(
-    (channel) => channel.type === 'email' && channel.isPrimary
-  )
-  if (!primaryContactChannel) {
-    logger.warn('No primary contact channel found for user', {
-      userPrimaryEmail: user.primaryEmail,
-      stackAuthUserId: user.id,
-    })
-    throw new Error('No primary contact channel found for user')
-  }
-  // @ts-expect-error - actually works but not correctly typed
-  await primaryContactChannel.sendVerificationEmail({
-    callbackUrl: core.billingPortalPageURL({
-      organizationId: params.organizationId,
-      customerExternalId: params.customerExternalId,
-      page: 'validate-magic-link',
-    }),
-  })
-  logger.info('Verification link email sent', {
-    userPrimaryEmail: user.primaryEmail,
-    stackAuthUserId: user.id,
-    primaryContactChannel: primaryContactChannel.value,
-  })
-}
-
-const createUserAndSendMagicLink = async (params: {
-  customerEmail: string
-  customerId: string
-  customerName: string
-  organizationId: string
-  customerExternalId: string
-}) => {
-  const tracer = trace.getTracer('magic-link')
-  return tracer.startActiveSpan(
-    'createUserAndSendMagicLink',
-    async (span) => {
-      try {
-        span.setAttributes({
-          'customer.email': params.customerEmail,
-          'customer.id': params.customerId,
-          'customer.name': params.customerName,
-        })
-
-        logger.info('Creating user and sending magic link', {
-          customerEmail: params.customerEmail,
-          customerId: params.customerId,
-          customerName: params.customerName,
-        })
-
-        const user = await hostedBillingStackServerApp.createUser({
-          primaryEmail: params.customerEmail,
-          primaryEmailAuthEnabled: true,
-          displayName: params.customerName,
-          otpAuthEnabled: true,
-          /**
-           * Todo: review this with Stack Auth team to ensure this is good to go
-           */
-          primaryEmailVerified: true,
-        })
-
-        await setHostedBillingCustomerExternalIdForStackAuthUser({
-          stackAuthUser: user,
-          organizationId: params.organizationId,
-          customerExternalId: params.customerExternalId,
-        })
-
-        logger.info('User created successfully', {
-          userId: user.id,
-          customerEmail: params.customerEmail,
-        })
-
-        await adminTransaction(async ({ transaction }) => {
-          await updateCustomer(
-            {
-              id: params.customerId,
-              stackAuthHostedBillingUserId: user.id,
-            },
-            transaction
-          )
-        })
-
-        logger.info('Customer updated with Stack Auth user ID', {
-          customerId: params.customerId,
-          stackAuthUserId: user.id,
-        })
-        await sendEmailToExistingUser({
-          user,
-          customerId: params.customerId,
-          customerName: params.customerName,
-          organizationId: params.organizationId,
-          customerExternalId: params.customerExternalId,
-        })
-        // keep this in case we need to go back to verify and magic link emails being different flows
-        // await sendVerificationEmailToUser(
-        //   user,
-        //   `${core.envVariable('HOSTED_BILLING_PORTAL_URL')}/api/${params.organizationId}/${params.customerExternalId}`
-        // )
-        span.setStatus({ code: SpanStatusCode.OK })
-        return { userId: user.id }
-      } catch (error) {
-        console.error('Error in createUserAndSendMagicLink', {
-          error,
-        })
-        span.setStatus({
-          code: SpanStatusCode.ERROR,
-          message:
-            error instanceof Error
-              ? error.message
-              : 'Failed to create user and send magic link',
-        })
-        span.setAttributes({ 'error.type': 'USER_CREATION_ERROR' })
-        logger.error('Failed to create user and send magic link', {
-          error:
-            error instanceof Error ? error.message : String(error),
-          customerEmail: params.customerEmail,
-          customerId: params.customerId,
-        })
-        throw error
-      } finally {
-        span.end()
-      }
-    }
-  )
-}
-
-async function sendEmailToExistingUser({
-  customerId,
-  user,
+async function sendMagicLinkAndUpdateCustomerForIdentifiedStackAuthUser({
+  customer,
   organizationId,
   customerExternalId,
 }: {
-  user: ServerUser
-  customerId: string
-  customerName: string
+  customer: Customer.Record
   organizationId: string
   customerExternalId: string
 }) {
-  logger.info(
-    'Customer already has Stack Auth user ID, sending magic link',
-    {
-      customerId,
-      stackAuthUserId: user.id,
+  if (!customer.stackAuthHostedBillingUserId) {
+    throw new Error('Customer has no stack auth user id')
+  }
+  const user = await hostedBillingStackServerApp.getUser(
+    customer.stackAuthHostedBillingUserId
+  )
+  if (!user) {
+    throw new Error('Stack Auth user not found')
+  }
+  await setHostedBillingCustomerExternalIdForStackAuthUser({
+    stackAuthUser: user,
+    organizationId,
+    customerExternalId,
+  })
+  await sendEmailToExistingUser({
+    user,
+    customerId: customer.id,
+    customerName: customer.name,
+    organizationId,
+    customerExternalId,
+  })
+}
+
+async function sendMagicLinkAndUpdateCustomerForUnidentifiedStackAuthUser({
+  customer,
+  organizationId,
+}: {
+  customer: Customer.Record
+  organizationId: string
+}) {
+  const stackAuthHostedBillingUserId = await adminTransaction(
+    async ({ transaction }) => {
+      return await mapCustomerEmailToStackAuthHostedBillingUserId(
+        customer.email,
+        transaction
+      )
     }
   )
-  const emailAndUserId = {
-    primaryEmail: user.primaryEmail!,
-    stackAuthUserId: user.id,
-    primaryEmailVerified: user.primaryEmailVerified,
-  }
-  if (user.primaryEmailVerified) {
-    await hostedBillingStackServerApp.sendMagicLinkEmail(
-      user.primaryEmail!,
-      {
-        callbackUrl: core.billingPortalPageURL({
-          organizationId,
-          customerExternalId,
-          page: 'validate-magic-link',
-        }),
-      }
-    )
-    logger.info(
-      'Primary email verified, login with magic link email sent to existing user',
-      emailAndUserId
-    )
-  } else {
-    await sendVerificationEmailToUser(user, {
+  if (!stackAuthHostedBillingUserId) {
+    await createUserAndSendMagicLink({
+      customerEmail: customer.email,
+      customerId: customer.id,
+      customerName: customer.name,
       organizationId,
-      customerExternalId,
+      customerExternalId: customer.externalId,
     })
-    logger.info(
-      'Primary email not verified, verification link email sent to existing user',
-      emailAndUserId
+  } else {
+    const user = await hostedBillingStackServerApp.getUser(
+      stackAuthHostedBillingUserId
     )
+    if (!user) {
+      throw new Error('Stack Auth user not found')
+    }
+    await sendEmailToExistingUser({
+      user,
+      customerId: customer.id,
+      customerName: customer.name,
+      organizationId,
+      customerExternalId: customer.externalId,
+    })
   }
 }
 
@@ -274,39 +165,25 @@ export const POST = withBillingApiRequestValidation(
               customerId: customer.id,
               customerEmail: customer.email,
             })
-            const user = await hostedBillingStackServerApp.getUser(
-              customer.stackAuthHostedBillingUserId
+            await sendMagicLinkAndUpdateCustomerForIdentifiedStackAuthUser(
+              {
+                customer,
+                organizationId,
+                customerExternalId,
+              }
             )
-            if (!user) {
-              throw new Error('Stack Auth user not found')
-            }
-            await setHostedBillingCustomerExternalIdForStackAuthUser({
-              stackAuthUser: user,
-              organizationId,
-              customerExternalId,
-            })
-            await sendEmailToExistingUser({
-              user,
-              customerId: customer.id,
-              customerName: customer.name,
-              organizationId,
-              customerExternalId,
-            })
           } else {
             logger.info('Creating new user and sending magic link', {
               customerId: customer.id,
               customerEmail: customer.email,
             })
-
-            await createUserAndSendMagicLink({
-              customerEmail,
-              customerId: customer.id,
-              customerName: customer.name,
-              organizationId,
-              customerExternalId,
-            })
+            await sendMagicLinkAndUpdateCustomerForUnidentifiedStackAuthUser(
+              {
+                customer,
+                organizationId,
+              }
+            )
           }
-
           span.setStatus({ code: SpanStatusCode.OK })
           return NextResponse.json({ success: true })
         } catch (error) {
