+++ platform/hosted-billing/src/app/(webapp)/layout.tsx
@@ -9,8 +9,8 @@ import { globalStackServerApp } from '@/stack'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Billing',
+  description: 'Billing',
 }
 
 export default function RootLayout({
